requirements.txt
asgiref==3.11.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
dj-rest-auth==7.0.1
Django==6.0
django-allauth==65.13.1
django-cors-headers==4.9.0
django-filter==25.2
django-ratelimit==4.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.2
Pillow==11.0.0
